Use PRIu32 in VADP dumper
Use PRIu32 for runtime meta-data length logging in the
VADP dumper instead of %u plus explicit casts.

+++ core/src/vmware/vadp_dumper/bareos_vadp_dumper.cc
@@ -25,6 +25,7 @@
  */
 #include <stdio.h>
 #include <stdint.h>
+#include <cinttypes>
 #include <string.h>
 #include <fcntl.h>
 #include <time.h>
@@ -1120,8 +1121,8 @@ static inline bool process_meta_data(bool validate_only)
 
     key = (char*)malloc(rmde.meta_key_length);
     if (!key) {
-      fprintf(stderr, "Failed to allocate %u bytes for meta data key\n",
-              static_cast<unsigned int>(rmde.meta_key_length));
+      fprintf(stderr, "Failed to allocate %" PRIu32 " bytes for meta data key\n",
+              rmde.meta_key_length);
       goto bail_out;
     }
 
@@ -1133,8 +1134,8 @@ static inline bool process_meta_data(bool validate_only)
 
     buffer = (char*)malloc(rmde.meta_data_length);
     if (!buffer) {
-      fprintf(stderr, "Failed to allocate %u bytes for meta data\n",
-              static_cast<unsigned int>(rmde.meta_data_length));
+      fprintf(stderr, "Failed to allocate %" PRIu32 " bytes for meta data\n",
+              rmde.meta_data_length);
       goto bail_out;
     }
